+++ src/modules/Home.js
@@ -3,17 +3,20 @@ import { StyleSheet, View } from 'react-native';
 import GetImage from './GetImage';
 
 const Home = () => {
-    const styles = StyleSheet.create({
-        container: {
-            flex: 1
-        }
-    });
     return (
         <View style={styles.container}>
-            <GetImage />
-            <StatusBar style="auto" />
+          <GetImage />
+          <StatusBar style="auto" />
         </View>
-    )
-}
+)};
 
-export default Home
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: '#fff',
+  }
+});
+
+export default Home;
